Moved analysis to their proper file location
