+++ src/analysis/playlistaData.ts
@@ -1,4 +1,4 @@
-import { Song } from './dataHandling.js';
+import { Song, artist_counter } from './dataHandling.js';
 import { spotify_get_request } from '../spotifyAPI.js';
 
 export async function get_playlist_content(playlistID: string, token: string) {
@@ -27,4 +27,10 @@ export async function get_playlist_content(playlistID: string, token: string) {
 	}
 
 	return playlist;
+}
+
+export async function analyze_playlist(id: string, token: string) {
+	const songList = await get_playlist_content(id, token);
+
+	return await artist_counter(songList)
 }
