+++ src/analysis/userData.ts
@@ -1,4 +1,4 @@
-import { Song, compare_songs } from './dataHandling.js';
+import { Song, artist_counter, compare_songs } from './dataHandling.js';
 import { get_playlist_content } from './playlistaData.js';
 import { spotify_get_request } from '../spotifyAPI.js';
 
@@ -36,4 +36,12 @@ export async function get_all_user_songs(playlists: string[], token: string) {
 	}
 
 	return all_songs;
+}
+
+export async function analayze_user(id: string, token: string) {
+	const userPlaylists = await get_user_playlists(id, token);
+
+	const songList = await get_all_user_songs(userPlaylists, token);
+
+	return await artist_counter(songList)
 }
